pyproject.toml
[project]
name = "local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.29.0",
    "fastapi>=0.128.0",
    "fastapi-mail>=1.6.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.9"
]
